update: ワークスペースの導入: coreモジュールに定義していた`Parser`をwasmモジュールに移動
併せて依存関係も整理した
