+++ core/Cargo.toml
@@ -13,18 +13,13 @@ categories.workspace = true
 [lib]
 crate-type = ["rlib", "cdylib"]
 
-[features]
-debug = []
-
 [dependencies]
-console_error_panic_hook = "0.1.7"
 itertools = "0.12.0"
 js-sys = "0.3.67"
 nom = "7.1.3"
 regex = "1.10.2"
 reqwest = { version = "0.11.23", default-features = false, features = ["json", "rustls-tls", "blocking"] }
 serde = { version = "1.0.192", features = ["derive"] }
-serde-wasm-bindgen = "0.6.1"
 wasm-bindgen = { workspace = true }
 wasm-bindgen-futures = { workspace = true }
