+++ wasm/Cargo.toml
@@ -11,3 +11,13 @@ categories.workspace = true
 
 [lib]
 crate-type = ["cdylib"]
+
+[features]
+debug = []
+
+[dependencies]
+console_error_panic_hook = "0.1.7"
+japanese-address-parser = { path = "../core" }
+serde-wasm-bindgen = "0.6.1"
+wasm-bindgen = { workspace = true }
+wasm-bindgen-futures = { workspace = true }
